Split tests by concern, tighten link types, unify navigation begin path
- Narrow To to string | NavigateTarget; new LinkTo type carries the
  link-only current override for useLinkProps/Link
- Funnel both navigation entry points through a single beginNavigation
  callback in Routes
- Split router.test.tsx into router/prepare/link/pending test files with
  shared helpers; flush jsdom's async fragment navigation inside act in
  the same-page hash test so stray router updates don't leak
